refactor(tui): split oversized app modules [OPE-123]
