+++ crates/opengoose-tui/src/app/credential_flow/tests.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+mod flow;
+mod persistence;
+mod provider_selection;
+mod support;
